refactor: route Sum-side toRepresentation through SumDispatch
Introduce SumDispatch and ProdDispatch as the entry-point dispatch layer
for representation conversions in the SIR lowerer. Sum-side
toRepresentation bodies (from SumCaseSirTypeGenerator,
SumCaseUplcConstrSirTypeGenerator, SumCaseUplcOnlySirTypeGenerator,
SumListCommonSirTypeGenerator) are lifted into SumDispatch.scala as
private helpers; each typegen's override now calls
SumDispatch.dispatcherBypass to surface any caller that bypasses the
dispatcher. LoweredValue.toRepresentation (and the VariableLoweredValue
/ DependendVariableLoweredValue overrides) always route through
SumDispatch — splitting by SIRType.isSum/isProd doesn't work because
sum-shaped constructors like List.Cons carry CaseClass SIR types whose
typegen is Sum-side.

Also removes three dead (DataConstr, _) arms in SumListCommon: the
DataConstr label means Data.Constr(tag, fields)-encoded bytes, but the
arms applied unListData (which only works on Data.List bytes). Probe
across the full test suite saw zero hits and the operation would have
failed at runtime anyway. ProdDispatch stays a pure forwarder until
Phase 2 lifts the Prod bodies the same way.
