+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumCaseSirTypeGenerator.scala
@@ -34,94 +34,8 @@ object SumCaseSirTypeGenerator extends SirTypeUplcGenerator {
         input: LoweredValue,
         representation: LoweredValueRepresentation,
         pos: SIRPosition
-    )(using lctx: LoweringContext): LoweredValue = {
-        (input.representation, representation) match {
-            case (DataConstr, DataConstr) =>
-                input
-            case (DataConstr, PairIntDataList) =>
-                lvBuiltinApply(SIRBuiltins.unConstrData, input, input.sirType, PairIntDataList, pos)
-            case (DataConstr, SumBuiltinList(elementRepr)) =>
-                // DataConstr → SumBuiltinList: go through PairIntDataList
-                input
-                    .toRepresentation(PairIntDataList, pos)
-                    .toRepresentation(SumBuiltinList(elementRepr), pos)
-            case (DataConstr, PackedSumDataList) =>
-                input
-                    .toRepresentation(PairIntDataList, pos)
-                    .toRepresentation(PackedSumDataList, pos)
-            case (DataConstr, SumUplcConstr(variants)) =>
-                // DataConstr → SumUplcConstr: go through PairIntDataList
-                input
-                    .toRepresentation(PairIntDataList, pos)
-                    .toRepresentation(representation, pos)
-            // === PairIntDataList → list reprs: delegate to list generator ===
-            case (PairIntDataList, SumBuiltinList(elementRepr)) =>
-                val elemType =
-                    SumBuiltinList.retrieveListElementType(input.sirType).getOrElse(SIRType.Data.tp)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                new SumBuiltinListSirTypeGenerator(elemRepr)
-                    .toRepresentation(input, representation, pos)
-            case (PairIntDataList, PackedSumDataList) =>
-                val elemType =
-                    SumBuiltinList.retrieveListElementType(input.sirType).getOrElse(SIRType.Data.tp)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                new SumBuiltinListSirTypeGenerator(elemRepr)
-                    .toRepresentation(input, representation, pos)
-            // PairIntDataList → SumUplcConstr: delegate
-            case (PairIntDataList, _: SumUplcConstr) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            // All SumBuiltinList source conversions delegate to SumBuiltinListSirTypeGenerator
-            case (SumBuiltinList(inElemRepr), _) =>
-                new SumBuiltinListSirTypeGenerator(inElemRepr)
-                    .toRepresentation(input, representation, pos)
-            case (PackedSumDataList, PackedSumDataList) =>
-                input
-            // All other PackedSumDataList source conversions delegate to SumBuiltinListSirTypeGenerator
-            case (PackedSumDataList, _) =>
-                val elemType =
-                    SumBuiltinList.retrieveListElementType(input.sirType).getOrElse(SIRType.Data.tp)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                new SumBuiltinListSirTypeGenerator(elemRepr)
-                    .toRepresentation(input, representation, pos)
-            // All SumUplcConstr source/target conversions delegate to SumUplcConstrSirTypeGenerator
-            case (_: SumUplcConstr, _) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            case (_, _: SumUplcConstr) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            case (_: ProductCaseClassRepresentation.ProdUplcConstr, _) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            case (inTvr: TypeVarRepresentation, outRepr) =>
-                import SIRType.TypeVarKind.*
-                inTvr.kind match
-                    case Transparent =>
-                        RepresentationProxyLoweredValue(input, representation, pos)
-                    case Unwrapped =>
-                        // Source bytes are in defaultRepresentation form. Relabel as that
-                        // underlying repr, then convert.
-                        val sourceUnderlying = defaultRepresentation(input.sirType)
-                        val r0 = RepresentationProxyLoweredValue(input, sourceUnderlying, pos)
-                        toRepresentation(r0, representation, pos)
-                    case Fixed =>
-                        val r0 = RepresentationProxyLoweredValue(input, DataConstr, pos)
-                        toRepresentation(r0, representation, pos)
-            case (inRepr, outTvr: TypeVarRepresentation) =>
-                import SIRType.TypeVarKind.*
-                outTvr.kind match
-                    case Transparent => input
-                    case Unwrapped   =>
-                        // Convert input to defaultRepresentation form, then relabel as Unwrapped.
-                        val targetUnderlying = defaultRepresentation(input.sirType)
-                        val converted = input.toRepresentation(targetUnderlying, pos)
-                        new RepresentationProxyLoweredValue(converted, outTvr, pos)
-                    case Fixed =>
-                        toRepresentation(input, DataConstr, pos)
-            case (_, _) =>
-                throw LoweringException(
-                  s"Unsupported conversion for ${input.sirType.show} from ${input.representation} to $representation",
-                  pos
-                )
-        }
-    }
+    )(using lctx: LoweringContext): LoweredValue =
+        SumDispatch.dispatcherBypass("SumCaseSirTypeGenerator")
 
     override def upcastOne(
         input: LoweredValue,
