+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumCaseUplcConstrSirTypeGenerator.scala
@@ -42,94 +42,8 @@ object SumCaseUplcConstrSirTypeGenerator extends SirTypeUplcGenerator {
         input: LoweredValue,
         representation: LoweredValueRepresentation,
         pos: SIRPosition
-    )(using lctx: LoweringContext): LoweredValue = {
-        // Dispatch based on input representation to avoid cycles:
-        // intermediates with SumBuiltinList repr should go through SumBuiltinList handler directly.
-        (input.representation, representation) match
-            // Identity
-            case (a, b) if a == b => input
-            case (a, b) if a.isCompatibleOn(input.sirType, b, pos) =>
-                RepresentationProxyLoweredValue(input, representation, pos)
-            // SumBuiltinList → DataConstr: listData(input) → PackedSumDataList → DataConstr
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(elemRepr),
-                  SumCaseClassRepresentation.DataConstr
-                ) =>
-                val asPackedList = new SumBuiltinListSirTypeGenerator(elemRepr)
-                    .toRepresentation(input, SumCaseClassRepresentation.PackedSumDataList, pos)
-                LoweredValue.Builder.lvBuiltinApply(
-                  SIRBuiltins.listData,
-                  asPackedList,
-                  input.sirType,
-                  SumCaseClassRepresentation.DataConstr,
-                  pos
-                )
-            // SumBuiltinList input → delegate to SumBuiltinList handler
-            case (SumCaseClassRepresentation.SumBuiltinList(elemRepr), _) =>
-                new SumBuiltinListSirTypeGenerator(elemRepr)
-                    .toRepresentation(input, representation, pos)
-            // PackedSumDataList input → delegate to SumBuiltinList handler
-            case (SumCaseClassRepresentation.PackedSumDataList, _) =>
-                new SumBuiltinListSirTypeGenerator(PrimitiveRepresentation.PackedData)
-                    .toRepresentation(input, representation, pos)
-            // DataConstr/PairIntDataList → SumBuiltinList/PackedSumDataList/PairIntDataList:
-            // delegate to SumCaseSirTypeGenerator which handles these without BuiltinList type hack
-            case (
-                  SumCaseClassRepresentation.DataConstr,
-                  _: SumCaseClassRepresentation.SumBuiltinList
-                ) | (
-                  SumCaseClassRepresentation.DataConstr,
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) | (
-                  SumCaseClassRepresentation.DataConstr,
-                  SumCaseClassRepresentation.PairIntDataList
-                ) | (
-                  SumCaseClassRepresentation.PairIntDataList,
-                  _: SumCaseClassRepresentation.SumBuiltinList
-                ) | (
-                  SumCaseClassRepresentation.PairIntDataList,
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                SumCaseSirTypeGenerator.toRepresentation(input, representation, pos)
-            // TypeVar source: dispatch by kind, then delegate or convert appropriately
-            case (inTvr: TypeVarRepresentation, _) =>
-                import SIRType.TypeVarKind.*
-                inTvr.kind match
-                    case Transparent =>
-                        // Wildcard source — delegate (existing behavior).
-                        SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-                    case Unwrapped =>
-                        // Source bytes are in defaultRepresentation form for input.sirType.
-                        val sourceUnderlying = defaultRepresentation(input.sirType)
-                        val r0 = RepresentationProxyLoweredValue(input, sourceUnderlying, pos)
-                        toRepresentation(r0, representation, pos)
-                    case Fixed =>
-                        // Source bytes are Data-shaped — delegate (existing behavior).
-                        SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            // TypeVar target: dispatch by kind
-            case (_, outTvr: TypeVarRepresentation) =>
-                import SIRType.TypeVarKind.*
-                outTvr.kind match
-                    case Transparent => input
-                    case Unwrapped =>
-                        val targetUnderlying = defaultRepresentation(input.sirType)
-                        val converted = input.toRepresentation(targetUnderlying, pos)
-                        new RepresentationProxyLoweredValue(converted, outTvr, pos)
-                    case Fixed =>
-                        val targetUnderlying = defaultTypeVarReperesentation(input.sirType)
-                        val converted = input.toRepresentation(targetUnderlying, pos)
-                        new RepresentationProxyLoweredValue(converted, outTvr, pos)
-            // SumUplcConstr, DataConstr, PairIntDataList → SumUplcConstrSirTypeGenerator
-            case (_: SumCaseClassRepresentation.SumUplcConstr, _) |
-                (SumCaseClassRepresentation.DataConstr, _) |
-                (SumCaseClassRepresentation.PairIntDataList, _) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, representation, pos)
-            case (inRepr, outRepr) =>
-                throw LoweringException(
-                  s"SumCaseUplcConstrSirTypeGenerator: unhandled conversion $inRepr → $outRepr for ${input.sirType.show}",
-                  pos
-                )
-    }
+    )(using lctx: LoweringContext): LoweredValue =
+        SumDispatch.dispatcherBypass("SumCaseUplcConstrSirTypeGenerator")
 
     override def upcastOne(input: LoweredValue, targetType: SIRType, pos: SIRPosition)(using
         lctx: LoweringContext
