+++ scalus-core/shared/src/main/scala/scalus/compiler/sir/lowering/typegens/SumListCommonSirTypeGenerator.scala
@@ -12,15 +12,6 @@ import scala.util.control.NonFatal
   */
 trait SumListCommonSirTypeGenerator extends SirTypeUplcGenerator {
 
-    /** Check if a TypeVarRepresentation is native based on its kind. */
-    protected def isNativeTypeVar(tvr: TypeVarRepresentation)(using
-        lctx: LoweringContext
-    ): Boolean =
-        import SIRType.TypeVarKind.*
-        tvr.kind match
-            case Transparent | Unwrapped => true
-            case Fixed                   => false
-
     def defaultListRepresentation(tp: SIRType, pos: SIRPosition)(using
         LoweringContext
     ): LoweredValueRepresentation
@@ -37,526 +28,12 @@ trait SumListCommonSirTypeGenerator extends SirTypeUplcGenerator {
 
     override def canBeConvertedToData(tp: SIRType)(using lctx: LoweringContext): Boolean = true
 
-    /** Convert a BuiltinList between different resolved element representations.
-      *
-      * Builds a conversion lambda and applies ScalusRuntime.mapList using lvApplyDirect.
-      *
-      * @param outputRepresentation
-      *   the original target representation from toRepresentation caller
-      */
-    private def convertBuiltinList(
-        input: LoweredValue,
-        elemType: SIRType,
-        resolvedIn: LoweredValueRepresentation,
-        resolvedOut: LoweredValueRepresentation,
-        outputRepresentation: LoweredValueRepresentation,
-        pos: SIRPosition
-    )(using lctx: LoweringContext): LoweredValue = {
-        val inElemUplcType = resolvedIn.uplcType(elemType)
-        val outElemUplcType = resolvedOut.uplcType(elemType)
-        val inListType = SIRType.BuiltinList(inElemUplcType)
-        val outListType = SIRType.BuiltinList(outElemUplcType)
-        val isPairIn = resolvedIn match
-            case _: ProductCaseClassRepresentation.ProdBuiltinPair => true
-            case _                                                 => false
-        val isPairOut = resolvedOut match
-            case _: ProductCaseClassRepresentation.ProdBuiltinPair => true
-            case _                                                 => false
-        val inListRepr: LoweredValueRepresentation =
-            if isPairIn then
-                SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-            else SumCaseClassRepresentation.SumBuiltinList(resolvedIn)
-        val outListRepr: LoweredValueRepresentation =
-            if isPairOut then
-                SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-            else SumCaseClassRepresentation.SumBuiltinList(resolvedOut)
-        val outNil =
-            if isPairOut then SumPairBuiltinListSirTypeGenerator.genNil(outListType, pos)
-            else new SumBuiltinListSirTypeGenerator(resolvedOut).genNil(outListType, pos)
-        val fnType = inElemUplcType ->: outElemUplcType
-        val fnRepr = LambdaRepresentation(fnType, InOutRepresentationPair(resolvedIn, resolvedOut))
-        val convFn = lvLamAbs(
-          "elem",
-          elemType,
-          resolvedIn,
-          elem => elem.toRepresentation(resolvedOut, pos),
-          pos
-        )
-        // mapList: (A->B) -> List[B] -> List[A] -> List[B]
-        val afterFnType = outListType ->: inListType ->: outListType
-        val afterFnRepr = LambdaRepresentation(
-          afterFnType,
-          InOutRepresentationPair(
-            outListRepr,
-            LambdaRepresentation(
-              inListType ->: outListType,
-              InOutRepresentationPair(inListRepr, outListRepr)
-            )
-          )
-        )
-        val afterNilType = inListType ->: outListType
-        val afterNilRepr =
-            LambdaRepresentation(afterNilType, InOutRepresentationPair(inListRepr, outListRepr))
-
-        val result = lvApplyDirect(
-          lvApplyDirect(
-            lvApplyDirect(
-              ScalusRuntime.mapList,
-              convFn,
-              afterFnType,
-              afterFnRepr,
-              pos
-            ),
-            outNil,
-            afterNilType,
-            afterNilRepr,
-            pos
-          ),
-          input,
-          input.sirType,
-          outListRepr,
-          pos
-        )
-        // If the caller's outputRepresentation differs from outListRepr, proxy it
-        if outputRepresentation == outListRepr then result
-        else RepresentationProxyLoweredValue(result, outputRepresentation, pos)
-    }
-
-    private def hasConstantOrTypeVar(
-        a: LoweredValueRepresentation,
-        b: LoweredValueRepresentation
-    ): Boolean =
-        a == PrimitiveRepresentation.Constant || b == PrimitiveRepresentation.Constant ||
-            a.isInstanceOf[TypeVarRepresentation] || b.isInstanceOf[TypeVarRepresentation]
-
     override def toRepresentation(
         input: LoweredValue,
         outputRepresentation: LoweredValueRepresentation,
         pos: SIRPosition
-    )(using lctx: LoweringContext): LoweredValue = {
-        if input.representation == outputRepresentation then return input
-        if input.representation.isCompatibleOn(input.sirType, outputRepresentation, pos) then
-            return RepresentationProxyLoweredValue(input, outputRepresentation, pos)
-        (input.representation, outputRepresentation) match
-            // === SumBuiltinList identity ===
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(inElemRepr),
-                  SumCaseClassRepresentation.SumBuiltinList(outElemRepr)
-                ) if inElemRepr == outElemRepr =>
-                input
-            // === SumBuiltinList cross-element-repr conversions ===
-            // Handle when at least one side involves Constant or TypeVarRepresentation.
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(inElemRepr),
-                  SumCaseClassRepresentation.SumBuiltinList(outElemRepr)
-                ) if hasConstantOrTypeVar(inElemRepr, outElemRepr) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                def resolveElementRepr(
-                    repr: LoweredValueRepresentation
-                ): LoweredValueRepresentation =
-                    repr match
-                        case tvr: TypeVarRepresentation =>
-                            val gen = lctx.typeGenerator(elemType)
-                            if !tvr.isPackedData then gen.defaultRepresentation(elemType)
-                            else gen.defaultTypeVarReperesentation(elemType)
-                        case other => other
-                val resolvedIn = resolveElementRepr(inElemRepr)
-                val resolvedOut = resolveElementRepr(outElemRepr)
-                val hasNativeTypeVar = (inElemRepr, outElemRepr) match
-                    case (tvr: TypeVarRepresentation, _) if isNativeTypeVar(tvr) => true
-                    case (_, tvr: TypeVarRepresentation) if isNativeTypeVar(tvr) => true
-                    case _                                                       => false
-                if resolvedIn == resolvedOut
-                    || elemType == SIRType.FreeUnificator
-                    || elemType == SIRType.TypeNothing
-                    || hasNativeTypeVar
-                then RepresentationProxyLoweredValue(input, outputRepresentation, pos)
-                else
-                    convertBuiltinList(
-                      input,
-                      elemType,
-                      resolvedIn,
-                      resolvedOut,
-                      outputRepresentation,
-                      pos
-                    )
-            // === SumBuiltinList(Constant) special cases ===
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(PrimitiveRepresentation.Constant),
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                input
-                    .toRepresentation(SumCaseClassRepresentation.SumBuiltinList(elemRepr), pos)
-                    .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-            case (
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  SumCaseClassRepresentation.SumBuiltinList(PrimitiveRepresentation.Constant)
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                input
-                    .toRepresentation(SumCaseClassRepresentation.SumBuiltinList(elemRepr), pos)
-                    .toRepresentation(outputRepresentation, pos)
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(PrimitiveRepresentation.Constant),
-                  tv @ TypeVarRepresentation(kind)
-                ) =>
-                import SIRType.TypeVarKind.*
-                kind match
-                    case Transparent =>
-                        // Wildcard target — relabel.
-                        new RepresentationProxyLoweredValue(input, tv, pos)
-                    case Unwrapped =>
-                        // Target Unwrapped wants bytes in defaultRepresentation form. For a
-                        // list of native primitives, source IS the default — pure relabel.
-                        new RepresentationProxyLoweredValue(input, tv, pos)
-                    case Fixed =>
-                        // Target Fixed wants Data-encoded bytes — convert via PackedSumDataList.
-                        input
-                            .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-                            .toRepresentation(tv, pos)
-            case (
-                  tv @ TypeVarRepresentation(kind),
-                  SumCaseClassRepresentation.SumBuiltinList(PrimitiveRepresentation.Constant)
-                ) =>
-                import SIRType.TypeVarKind.*
-                kind match
-                    case Transparent =>
-                        // Wildcard source — relabel as native list.
-                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
-                    case Unwrapped =>
-                        // Source bytes already in defaultRepresentation form (= target).
-                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
-                    case Fixed =>
-                        // Source is Data-encoded — convert via PackedSumDataList.
-                        input
-                            .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-                            .toRepresentation(outputRepresentation, pos)
-            // === SumBuiltinList → PackedSumDataList (via listData) ===
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(_),
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val asDataList =
-                    if elemType == SIRType.TypeNothing || elemType == SIRType.FreeUnificator then
-                        input
-                    else
-                        val elemRepr =
-                            lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                        input.toRepresentation(
-                          SumCaseClassRepresentation.SumBuiltinList(elemRepr),
-                          pos
-                        )
-                lvBuiltinApply(
-                  SIRBuiltins.listData,
-                  asDataList,
-                  input.sirType,
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  pos
-                )
-            // === SumBuiltinList catch-all (compatible element reprs) ===
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(inElemRepr),
-                  out @ SumCaseClassRepresentation.SumBuiltinList(outElemRepr)
-                ) =>
-                if inElemRepr == outElemRepr then input
-                else if (inElemRepr == PrimitiveRepresentation.PackedData && outElemRepr == SumCaseClassRepresentation.DataData)
-                    || (inElemRepr == SumCaseClassRepresentation.DataData && outElemRepr == PrimitiveRepresentation.PackedData)
-                then
-                    throw LoweringException(
-                      s"PackedData/DataData mismatch in SumBuiltinList: ${inElemRepr} -> ${outElemRepr} " +
-                          s"type=${input.sirType.show} createdEx=${input.createdEx}",
-                      pos
-                    )
-                else
-                    val elemType = retrieveElementType(input.sirType, pos)
-                    if inElemRepr.isCompatibleOn(elemType, outElemRepr, pos) then
-                        RepresentationProxyLoweredValue(input, out, pos)
-                    else convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
-            // === SumBuiltinList → SumDataAssocMap (go through SumPairBuiltinList) ===
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(_),
-                  SumCaseClassRepresentation.SumDataAssocMap
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val pairRepr =
-                    SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-                input
-                    .toRepresentation(pairRepr, pos)
-                    .toRepresentation(SumCaseClassRepresentation.SumDataAssocMap, pos)
-            // === PackedSumDataList conversions ===
-            case (
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  out @ SumCaseClassRepresentation.SumBuiltinList(_)
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                val dataListRepr = SumCaseClassRepresentation.SumBuiltinList(elemRepr)
-                val asDataList = lvBuiltinApply(
-                  SIRBuiltins.unListData,
-                  input,
-                  input.sirType,
-                  dataListRepr,
-                  pos
-                )
-                if out == dataListRepr then asDataList
-                else asDataList.toRepresentation(out, pos)
-            case (
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                input
-            case (
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  SumCaseClassRepresentation.SumDataAssocMap
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                val pairRepr =
-                    SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-                input
-                    .toRepresentation(SumCaseClassRepresentation.SumBuiltinList(elemRepr), pos)
-                    .toRepresentation(pairRepr, pos)
-                    .toRepresentation(SumCaseClassRepresentation.SumDataAssocMap, pos)
-            case (
-                  SumCaseClassRepresentation.PackedSumDataList,
-                  out @ SumCaseClassRepresentation.SumPairBuiltinList(_, _)
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                input
-                    .toRepresentation(SumCaseClassRepresentation.SumBuiltinList(elemRepr), pos)
-                    .toRepresentation(out, pos)
-            // === SumPairBuiltinList conversions ===
-            case (
-                  SumCaseClassRepresentation.SumPairBuiltinList(_, _),
-                  SumCaseClassRepresentation.SumPairBuiltinList(_, _)
-                ) =>
-                input
-            case (
-                  SumCaseClassRepresentation.SumPairBuiltinList(_, _),
-                  SumCaseClassRepresentation.SumDataAssocMap
-                ) =>
-                lvBuiltinApply(
-                  SIRBuiltins.mapData,
-                  input,
-                  input.sirType,
-                  SumCaseClassRepresentation.SumDataAssocMap,
-                  pos
-                )
-            case (
-                  SumCaseClassRepresentation.SumPairBuiltinList(_, _),
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                // PackedSumDataList is List-shaped Data, NOT Map-shaped — so go via
-                // SumBuiltinList(ProdDataConstr) → listData. The previous path through
-                // SumDataAssocMap (mapData) packed as Map Data, which is wrong AND looped
-                // with line 371's (SumDataAssocMap, _) catch-all that re-emits via
-                // SumPairBuiltinList.
-                val elemType = retrieveElementType(input.sirType, pos)
-                val dataElemRepr =
-                    lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                input
-                    .toRepresentation(SumCaseClassRepresentation.SumBuiltinList(dataElemRepr), pos)
-                    .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-            case (
-                  SumCaseClassRepresentation.SumPairBuiltinList(inKeyRepr, inValueRepr),
-                  out @ SumCaseClassRepresentation.SumBuiltinList(outElemRepr)
-                ) =>
-                // PairList → regular list: element-wise conversion via convertBuiltinList
-                val elemType = retrieveElementType(input.sirType, pos)
-                val inElemRepr =
-                    ProductCaseClassRepresentation.ProdBuiltinPair(inKeyRepr, inValueRepr)
-                convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
-            case (
-                  SumCaseClassRepresentation.SumBuiltinList(inElemRepr),
-                  out @ SumCaseClassRepresentation.SumPairBuiltinList(outKeyRepr, outValueRepr)
-                ) =>
-                // Regular list → PairList: element-wise conversion via convertBuiltinList
-                if isNilType(input.sirType) then lvPairDataNil(pos, input.sirType, out)
-                else
-                    val elemType = retrieveElementType(input.sirType, pos)
-                    val outElemRepr =
-                        ProductCaseClassRepresentation.ProdBuiltinPair(outKeyRepr, outValueRepr)
-                    convertBuiltinList(input, elemType, inElemRepr, outElemRepr, out, pos)
-            // === SumDataAssocMap conversions ===
-            case (
-                  SumCaseClassRepresentation.SumDataAssocMap,
-                  SumCaseClassRepresentation.SumPairBuiltinList(_, _)
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val pairRepr =
-                    SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-                lvBuiltinApply(
-                  SIRBuiltins.unMapData,
-                  input,
-                  input.sirType,
-                  pairRepr,
-                  pos
-                )
-            case (SumCaseClassRepresentation.SumDataAssocMap, _) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val pairRepr =
-                    SumCaseClassRepresentation.SumPairBuiltinList.fromElementType(elemType, pos)
-                input
-                    .toRepresentation(pairRepr, pos)
-                    .toRepresentation(outputRepresentation, pos)
-            // === DataConstr → SumUplcConstr: unListData → builtinListToUplcConstr ===
-            case (
-                  SumCaseClassRepresentation.DataConstr,
-                  outSum: SumCaseClassRepresentation.SumUplcConstr
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                val dataListRepr = SumCaseClassRepresentation.SumBuiltinList(elemRepr)
-                val asBuiltinList = lvBuiltinApply(
-                  SIRBuiltins.unListData,
-                  input,
-                  input.sirType,
-                  dataListRepr,
-                  pos
-                )
-                ScalusRuntime.builtinListToUplcConstr(asBuiltinList, outSum, input.sirType, pos)
-            // === DataConstr → PackedSumDataList: direct via listData ===
-            case (
-                  SumCaseClassRepresentation.DataConstr,
-                  SumCaseClassRepresentation.PackedSumDataList
-                ) =>
-                val elemType = retrieveElementType(input.sirType, pos)
-                val elemRepr = lctx.typeGenerator(elemType).defaultDataRepresentation(elemType)
-                val dataListRepr = SumCaseClassRepresentation.SumBuiltinList(elemRepr)
-                val asDataList = lvBuiltinApply(
-                  SIRBuiltins.unListData,
-                  input,
-                  input.sirType,
-                  dataListRepr,
-                  pos
-                )
-                if dataListRepr == outputRepresentation then asDataList
-                else asDataList.toRepresentation(outputRepresentation, pos)
-            // === DataConstr → other list repr: go through PackedSumDataList ===
-            case (SumCaseClassRepresentation.DataConstr, _) =>
-                input
-                    .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-                    .toRepresentation(outputRepresentation, pos)
-            // === PairIntDataList → list repr: go through DataConstr → PackedSumDataList ===
-            case (
-                  SumCaseClassRepresentation.PairIntDataList,
-                  SumCaseClassRepresentation.SumBuiltinList(elementRepr)
-                ) =>
-                // TODO: rewrie without going through DataConstr, by directly converting the (tag, fieldList) pair to a builtin list of the right element repr.
-                val tag = lvBuiltinApply(
-                  SIRBuiltins.fstPair,
-                  input,
-                  SIRType.Integer,
-                  PrimitiveRepresentation.Constant,
-                  pos
-                )
-                // If tag == 0, it's a Nil, otherwise it's a Cons with one element (the fieldList).
-                throw LoweringException(
-                  s"PairIntDataList → SumBuiltinList($elementRepr) for ${input.sirType.show}: " +
-                      "DataConstr should not exist for List types.",
-                  pos
-                )
-            case (SumCaseClassRepresentation.PairIntDataList, _) =>
-                // PairIntDataList is (tag, fieldList) from unConstrData.
-                // For lists, reconstruct DataConstr, then go through PackedSumDataList.
-                val asDataConstr = lvBuiltinApply2(
-                  SIRBuiltins.constrData,
-                  lvBuiltinApply(
-                    SIRBuiltins.fstPair,
-                    input,
-                    SIRType.Integer,
-                    PrimitiveRepresentation.Constant,
-                    pos
-                  ),
-                  lvBuiltinApply(
-                    SIRBuiltins.sndPair,
-                    input,
-                    SIRType.List(SIRType.Data.tp),
-                    SumCaseClassRepresentation.SumBuiltinList(SumCaseClassRepresentation.DataData),
-                    pos
-                  ),
-                  input.sirType,
-                  SumCaseClassRepresentation.DataConstr,
-                  pos
-                )
-                asDataConstr
-                    .toRepresentation(SumCaseClassRepresentation.PackedSumDataList, pos)
-                    .toRepresentation(outputRepresentation, pos)
-            // === TypeVarRepresentation ===
-            case (_, tv: TypeVarRepresentation) =>
-                import SIRType.TypeVarKind.*
-                tv.kind match
-                    case Transparent =>
-                        // Wildcard target — relabel.
-                        new RepresentationProxyLoweredValue(input, tv, pos)
-                    case Unwrapped =>
-                        // Convert input to defaultRepresentation form, then relabel.
-                        val targetUnderlying = defaultRepresentation(input.sirType)
-                        val converted = input.toRepresentation(targetUnderlying, pos)
-                        new RepresentationProxyLoweredValue(converted, tv, pos)
-                    case Fixed =>
-                        val inputAsData = input.toRepresentation(
-                          SumCaseClassRepresentation.PackedSumDataList,
-                          pos
-                        )
-                        new RepresentationProxyLoweredValue(inputAsData, tv, pos)
-            case (tv: TypeVarRepresentation, _) =>
-                import SIRType.TypeVarKind.*
-                tv.kind match
-                    case Transparent =>
-                        new RepresentationProxyLoweredValue(input, outputRepresentation, pos)
-                    case Unwrapped =>
-                        // Source bytes are in defaultRepresentation form. Relabel as that
-                        // underlying repr, then convert.
-                        val sourceUnderlying = defaultRepresentation(input.sirType)
-                        val r0 = new RepresentationProxyLoweredValue(input, sourceUnderlying, pos)
-                        r0.toRepresentation(outputRepresentation, pos)
-                    case Fixed =>
-                        if input.representation == outputRepresentation then input
-                        else
-                            val r0 = new RepresentationProxyLoweredValue(
-                              input,
-                              SumCaseClassRepresentation.PackedSumDataList,
-                              pos
-                            )
-                            r0.toRepresentation(outputRepresentation, pos)
-            // SumReprProxy: unwrap and delegate
-            case (_: SumCaseClassRepresentation.SumReprProxy, _) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
-            case (_, _: SumCaseClassRepresentation.SumReprProxy) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
-            // SumUplcConstr → anything: delegate to SumUplcConstrSirTypeGenerator
-            case (_: SumCaseClassRepresentation.SumUplcConstr, _) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
-            // anything → SumUplcConstr: delegate to SumUplcConstrSirTypeGenerator
-            case (_, _: SumCaseClassRepresentation.SumUplcConstr) =>
-                SumUplcConstrSirTypeGenerator.toRepresentation(input, outputRepresentation, pos)
-            // ProdUplcConstr value at a sum-list-type site: wrap as a singleton
-            // SumUplcConstr of the input's variant, then delegate. This happens when a
-            // variant built via native Constr emission (e.g. `List.Nil` inside a
-            // native-Constr dispatcher scope) flows into code expecting the sum form.
-            case (puc: ProductCaseClassRepresentation.ProdUplcConstr, _) =>
-                val wrappedRepr =
-                    SumCaseClassRepresentation.SumUplcConstr(
-                      scala.collection.immutable.SortedMap(puc.tag -> puc)
-                    )
-                val wrapped = new RepresentationProxyLoweredValue(input, wrappedRepr, pos)
-                SumUplcConstrSirTypeGenerator.toRepresentation(
-                  wrapped,
-                  outputRepresentation,
-                  pos
-                )
-            case _ =>
-                throw LoweringException(
-                  s"Unexpected representation conversion for ${input.sirType.show} from ${input.representation} to ${outputRepresentation}",
-                  pos
-                )
-    }
+    )(using lctx: LoweringContext): LoweredValue =
+        SumDispatch.dispatcherBypass("SumListCommonSirTypeGenerator")
 
     override def upcastOne(
         input: LoweredValue,
